+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Attempting to start recording while recording is in progress will now raise a
+  RuntimeError.
 - [#77] Have the Django integration capture parameters of `http_server_request`s,
   recording them in the `message` attribute of the `call` event.
 - [#83] Capture HTTP request headers in django and flask.